scripts/recommendation_daily/requirements.txt
numpy>=1.26
pandas>=2.2
Pillow>=10.2
PyMySQL>=1.1
tqdm>=4.66
torch>=2.2
torchvision>=0.20
transformers>=4.57,<5
safetensors>=0.4
timm>=1.0
requests>=2.32
einops>=0.8
sentencepiece>=0.2
protobuf>=5
